backend/requirements.txt
# FastAPI & Server
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# Database
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
alembic==1.15.2

# AI / NLP
google-generativeai==0.8.5

# Document Processing
pdfplumber==0.11.6
python-docx==1.1.2
Pillow==11.2.1
pytesseract==0.3.13
openpyxl==3.1.5
python-pptx==1.0.2

# Utilities
python-dotenv==1.1.0
pydantic[email]==2.11.3
pydantic-settings==2.9.1
simhash==2.1.2
aiofiles==24.1.0
httpx==0.28.1

# CORS
starlette==0.46.2

# For mock data generation
fpdf2==2.8.3
Faker==37.1.0
